lib/game: Implement Graphic2D, for drawing with a coordinate system, which is independent of the actual screen

+++ src/lib/util/graphic/LineDrawer.cpp
@@ -22,7 +22,7 @@ namespace Util::Graphic {
 
 LineDrawer::LineDrawer(const PixelDrawer &pixelDrawer) : pixelDrawer(pixelDrawer) {}
 
-void LineDrawer::drawLine(uint16_t x1, uint16_t y1, uint16_t x2, uint16_t y2, const Color &color) const {
+void LineDrawer::drawLine(int32_t x1, int32_t y1, int32_t x2, int32_t y2, const Color &color) const {
     // Calculate our deltas
     auto dx = static_cast<int32_t>(x2 - x1);
     auto dy = static_cast<int32_t>(y2 - y1);
@@ -65,7 +65,7 @@ void LineDrawer::drawLine(uint16_t x1, uint16_t y1, uint16_t x2, uint16_t y2, co
     }
 }
 
-void LineDrawer::drawLineMajorAxis(uint16_t x, uint16_t y, int8_t xMovement, int8_t yMovement, int32_t dx, int32_t dy, bool majorAxisX, const Color &color) const {
+void LineDrawer::drawLineMajorAxis(int32_t x, int32_t y, int8_t xMovement, int8_t yMovement, int32_t dx, int32_t dy, bool majorAxisX, const Color &color) const {
     // Calculate some constants
     const int32_t dx2 = dx * 2;
     const int32_t dy2 = dy * 2;
@@ -100,7 +100,7 @@ void LineDrawer::drawLineMajorAxis(uint16_t x, uint16_t y, int8_t xMovement, int
     }
 }
 
-void LineDrawer::drawLineSingleAxis(uint16_t x, uint16_t y, int8_t movement, int32_t dx, bool majorAxisX, const Color &color) const {
+void LineDrawer::drawLineSingleAxis(int32_t x, int32_t y, int8_t movement, int32_t dx, bool majorAxisX, const Color &color) const {
     // Draw the first pixel
     pixelDrawer.drawPixel(x, y, color);
 
@@ -117,7 +117,7 @@ void LineDrawer::drawLineSingleAxis(uint16_t x, uint16_t y, int8_t movement, int
 
 }
 
-void LineDrawer::swap(uint16_t *a, uint16_t *b) {
+void LineDrawer::swap(int32_t *a, int32_t *b) {
     uint32_t h = *a;
     *a = *b;
     *b = h;
